Dynamically run jobs based on file extension
This commit changes the default functionality of Spring Batch - a Job is still ran at the start of the application, but now it applies custom logic to choose the correct Job.
An `app.filePath` property is now expected to be supplied at application start time, and no default is provided.
If the application fails to determine the Job to run, it will just run without running a job.

New dependencies:
- commons-io to get file extension